Refactored functional tests
